fix(game): fix problem in how production output effects were applied.
before they were applied *after* production had been calculated. for
example, say there's a 10% reduction in plant calories. with plant
calorie demand at 100, we'd produce 100, and *then* apply the reduction
so that the "actual" production amount is just 90, thus leading to a
shortage (90 < 100).

the changed version just updates the per-process output modifier,
which is what I should have been doing from the start :)

likely the cause of #298 and #280